+++ src/Model/LogoFunctions.java
@@ -2,6 +2,6 @@ package Model;
 
 public interface LogoFunctions {
 
-		public int interpret();
+		public double interpret();
 		
 }
